overlays: wm8960-soundcard: Fix clock declaration
How did this ever work? The static-clock declaration is fine, but the
reference to it is attached to part of the simple-audio-card node,
rather than the instantiation of the codec driver (a subnode of the
I2C controller).

Move the clock reference where it should be, and fix a few minor
cosmetic issues at the same time.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>